+++ packages/local_auth/local_auth_android/lib/local_auth_android.dart
@@ -55,20 +55,17 @@ class LocalAuthAndroid extends LocalAuthPlatform {
   @override
   Future<List<BiometricType>> getEnrolledBiometrics() async {
     final List<String> result = (await _channel.invokeListMethod<String>(
-          'getAvailableBiometrics',
+          'getEnrolledBiometrics',
         )) ??
         <String>[];
     final List<BiometricType> biometrics = <BiometricType>[];
     for (final String value in result) {
       switch (value) {
-        case 'face':
-          biometrics.add(BiometricType.face);
+        case 'weak':
+          biometrics.add(BiometricType.weak);
           break;
-        case 'fingerprint':
-          biometrics.add(BiometricType.fingerprint);
-          break;
-        case 'iris':
-          biometrics.add(BiometricType.iris);
+        case 'strong':
+          biometrics.add(BiometricType.strong);
           break;
         case 'undefined':
           break;