+++ packages/local_auth/local_auth_android/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+* Fix getEnrolledBiometrics to be in line with iOS implementation:
+  Present biometrics that are not enrolled are no longer returned.
+
 ## 1.0.0
 
 * Initial release from migration to federated architecture.